chore: order result component
